+++ requirements.txt
@@ -2,4 +2,5 @@ colorama
 pwinput
 psutil
 art
-lark
+#lark
+#ply
